added nodejs rendering
